Alterado tipo de dados da colunas

+++ src/br/com/constants/HasDocument.java
@@ -2,8 +2,8 @@ package br.com.constants;
 
 public interface HasDocument {
 
-    public Integer getDocument();
+    public Long getDocument();
 
-    public void setDocument(Integer document);
+    public void setDocument(Long document);
 
 }
