+++ src/br/com/entitys/Client.java
@@ -64,7 +64,7 @@ public class Client extends Entity implements HasAddress, HasUser, HasContact, H
     private String email;
 
     @Column(name = Client.Columns.PHONE, isBigInt = true, unique = true)
-    private Integer phone;
+    private Long phone;
 
     @Column(name = Client.Columns.PHONE_TYPE, isText = true)
     private PhoneType phoneType;
@@ -73,7 +73,7 @@ public class Client extends Entity implements HasAddress, HasUser, HasContact, H
     private PersonType personType;
 
     @Column(name = Client.Columns.DOCUMENT, isBigInt = true, unique = true)
-    private Integer document;
+    private Long document;
 
     @Column(name = Client.Columns.IS_ACTIVE, notNull = true, defaultValue = "true")
     private Boolean isActive;
@@ -142,12 +142,12 @@ public class Client extends Entity implements HasAddress, HasUser, HasContact, H
     }
 
     @Override
-    public Integer getPhone() {
+    public Long getPhone() {
         return phone;
     }
 
     @Override
-    public void setPhone(Integer phone) {
+    public void setPhone(Long phone) {
         this.phone = phone;
     }
 
@@ -172,12 +172,12 @@ public class Client extends Entity implements HasAddress, HasUser, HasContact, H
     }
 
     @Override
-    public Integer getDocument() {
+    public Long getDocument() {
         return document;
     }
 
     @Override
-    public void setDocument(Integer document) {
+    public void setDocument(Long document) {
         this.document = document;
     }
